Reorganize _runners.cjs and imports in fuzz targets
