+++ test/fuzz/_.cjs
@@ -1,0 +1,12 @@
+/**
+ * @overview Provides fuzzing utilities.
+ * @license MIT
+ */
+
+const common = require("./_common.cjs");
+const runners = require("../_runners.cjs");
+
+module.exports = {
+  common,
+  runners,
+};
